Add conpty backend
There is a bug in original conpty_terminate function.
If session is closed from local end first it will hang.
Pterm simply doesn't close backend at exit, so there this issue was not found.
The fix is to move ClosePseudoConsole() call to the beginning.
Because this fix the original conpty.c can't be used.