Simplify scrollbar update at session activation
Earlier the scroll info was stored in WinGuiFrontend.
At activation it was just used to set the scroll bar or if
term->win_scrollbar_update_pending was true recalculation happened.
Since recalculation is easy, now we do it unconditionally.
No need to store scroll info also we don't use on term internal variable.